final_project/tests_app/requirements.txt
pytest==7.1.3
selenium==4.5.0
webdriver-manager==3.8.3
SQLAlchemy==1.4.44
pytest-xdist==2.5.0
PyMySQL==1.0.2
cryptography==38.0.3
requests==2.28.1
allure-pytest==2.11.1
filelock==3.8.0